Permissions, Roles & Gates on the BE

+++ app/Http/Controllers/Api/PostController.php
@@ -46,6 +46,7 @@ class PostController extends Controller
 
     public function store(StorePostRequest $request)
     {
+        $this->authorize('posts.create');
         if ($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
             info($filename);
@@ -56,11 +57,13 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
+        $this->authorize('posts.update');
         return new PostResource($post);
     }
 
     public function update(Post $post, StorePostRequest $request)
     {
+        $this->authorize('posts.update');
         // if ($request->hasFile('thumbnail')) {
         //     $filename = $request->file('thumbnail')->getClientOriginalName();
         //     info($filename);
@@ -71,6 +74,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        // if logged in as a user w/out that permission, 403 unauthorized will be thrown
+        $this->authorize('posts.delete');
         $post->delete();
         return response()->noContent();
     }
